refactor: extract Picker out of input.ts into picker.ts
Standalone functions pick(), pickMultiple(), pickQuestion(), and promptLine()
have zero dependency on the Input class or Display, so they live in their own
picker.ts module. Input now delegates to these functions via thin wrappers.
Types PickConfig, PickResult, PickQuestionResult move to picker.ts; callers
updated to import from the new module location.

Closes #724
