+++ CLAUDE.md
@@ -22,7 +22,7 @@ A unified REPL + worker loop. `AgentController` in `controllers/agent-controller
 Follows MVC with three subdirectories:
 
 - **Models** (`models/`) — `workspace.ts` (git/npm workspace management; extends `EventEmitter`, emits `"clone-start"`, `"reset-start"`, `"reset-retry"`, `"reset-reclone"`, `"npm-install"`, `"destroy"`, `"prune-start"`, `"prune-remove"`, `"prune-skip"` events instead of calling `print()` directly — callers subscribe in `startWorkerMode()`; `prune()` is an instance method, not static), `settings.ts` (pure model for runtime-settable preferences; `Settings` is the only value export — owns model and effort state, emits `"change"` on updates via `_setModel`/`_setEffort`; effort/model constants and cache helpers (`EFFORT_LEVELS`, `VALID_EFFORT_VALUES`, `getCachedModels`, `setCachedModels`, `_resetCachedModels`, `findModel`) are all static members of `Settings` — no interactive methods), `query-stats.ts` (tracks token usage and turn counts; exports `QueryStats` class only — `WORKING_VERBS` and `pickWorkingVerb` are private to the module; no view imports)
-- **Views** (`views/`) — `display.ts` (TUI rendering; exports `Display` class with `BrunelConfig` and `StatusBar` injected in constructor — class methods use `this.config.verbose` and `this.statusBar`; `display.clearBreak()` and `display.fmtArgs(input)` delegate to module-level utilities; all pure utilities like `c`, `s`, `resolve`, `fmtEvent`, etc.; no singletons or standalone delegates), `status-bar.ts` (status bar state and rendering; exports `StatusBar` class only — holds all worker status state; constructed in `index.ts` startup and injected into `Display` and `startWorkerMode`; no singleton or `initStatusBar` export), `input.ts` (readline-based REPL input; exports `Input` class with `display: Display` injected in constructor — all input logic lives directly in class methods: `input.ask(promptStr, getCommands?, abort?)`, `input.pick(options, config?)`, `input.pickMultiple(options, promptStr?)`, `input.pickQuestion(options)`, `input.promptLine(prompt)`; also exports `_resetStash` for testing and `CommandSuggestion`/`PickConfig`/`PickResult`/`PickQuestionResult` types; no module-level standalone functions)
+- **Views** (`views/`) — `display.ts` (TUI rendering; exports `Display` class with `BrunelConfig` and `StatusBar` injected in constructor — class methods use `this.config.verbose` and `this.statusBar`; `display.clearBreak()` and `display.fmtArgs(input)` delegate to module-level utilities; all pure utilities like `c`, `s`, `resolve`, `fmtEvent`, etc.; no singletons or standalone delegates), `status-bar.ts` (status bar state and rendering; exports `StatusBar` class only — holds all worker status state; constructed in `index.ts` startup and injected into `Display` and `startWorkerMode`; no singleton or `initStatusBar` export), `input.ts` (readline-based REPL input; exports `Input` class with `display: Display` injected in constructor — `input.ask(promptStr, getCommands?, abort?)` is the core readline method with bracketed paste, multiline, and status bar integration; `input.pick`, `input.pickMultiple`, `input.pickQuestion`, `input.promptLine` are thin delegates to the standalone functions in `picker.ts`; also exports `_resetStash` for testing), `picker.ts` (standalone arrow-key picker functions with no dependency on `Input` or `Display`; exports `pick(options, config?)`, `pickMultiple(options, promptStr?)`, `pickQuestion(options)`, `promptLine(prompt)`, and types `PickConfig`, `PickResult`, `PickQuestionResult`)
 - **Controllers** (`controllers/`) — `agent-controller.ts` (`AgentController` class — handles the "run a query" action; constructor takes `(display, input, permConfig, settings)`; `runQuery(prompt, sessionId, ac?, model?, effort?)` runs a single turn via the SDK (handles `AskUserQuestion`/tool-permission callbacks, session ID capture, status bar lifecycle, ^C interrupt); also exports `logFull` (appends structured JSON to `repl.log`) and `createFetchModelsFn(permConfig)` (returns a function that fetches available models from the SDK)), `worker-controller.ts` (WS protocol + task lifecycle; defines `WorkerDisplay` interface `{ print, printForemanMessage }` — used by all controllers; owns `WorkerSession` with `session.confirmTaskQuit(info, pickFn?)` as a method and `WorkerSession.generateAgentId()` as a static method; exports `startWorkerMode(display, statusBar, input)` and `registerWorkerCommands(session, registry, display)`), `workspace-controller.ts` (registers `/workspace:*` slash commands; `registerWorkspaceCommands(workspace, registry, display)` takes injected `display: WorkerDisplay`), `command-controller.ts` (`CommandRegistry` class — registration only (`register`, `scoped`, `lookup`, `listAll`, `execute`, `_reset`); `CommandController` class — has-a `CommandRegistry` (not extends), adds `dispatch`, `parseSlashCommand`, `listCommandNames`, `listCommands`, `suggest`; exposes `controller.registry` getter for registration in tests; `index.ts` creates `new CommandRegistry()` for registration and `new CommandController(registry)` for dispatch; pass `CommandRegistry` to functions that only register commands), `settings-controller.ts` (`SettingsController` class — wraps `Settings` model with `display: WorkerDisplay`; exposes `pickModel(args, pickFn, fetchModelsFn)` and `pickEffort(args, pickFn)`)
 
 Other key files: `index.ts` (composition root + routing loop; exports `main(runQueryFn, permConfig, display, settings, input, runWorkerMode?, workspaceCfg?)` which sets up worker mode/workspace, prints the startup banner, registers all commands, and runs the routing loop — dispatching commands to their handlers, user prompts to `runQueryFn` i.e. `AgentController.runQuery`, and foreman signals to `WorkerSession`; the startup block at the bottom constructs objects in order (`settings` → `statusBar` → `display` → `new Input(display)`) and calls `main()`), `worker-prompts.ts` (prompt templates).
